refactoring package for aboutApp navigation and do reformat code for all project
