+++ mockApiServer/src/main/java/ru/sibsutis/mockapiserver/assets/AssetAdapter.kt
@@ -3,14 +3,14 @@ package ru.sibsutis.mockapiserver.assets
 import android.content.Context
 import android.net.Uri
 import android.util.Log
+import java.io.InputStreamReader
+import java.net.HttpURLConnection.HTTP_NOT_FOUND
+import java.net.HttpURLConnection.HTTP_OK
 import okhttp3.Interceptor
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.Response
 import okhttp3.ResponseBody.Companion.toResponseBody
 import okio.Buffer
-import java.io.InputStreamReader
-import java.net.HttpURLConnection.HTTP_NOT_FOUND
-import java.net.HttpURLConnection.HTTP_OK
 
 internal fun getFake(context: Context, uri: Uri, response: Response.Builder): Response.Builder =
 	when (uri.path) {
